오류로 TermOfUseBtn 컴포넌트 제거
