--- src/components/Register/TermsOfUse/TermOfUseList/TermsOfUseBtn/index.tsx
@@ -1,10 +1,0 @@
-import { AccordionBtn, TermsOfUseIcon } from './styles'
-
-const TermsOfUseBtn = ({ onAccordion }: any) => {
-  return (
-    <AccordionBtn onClick={onAccordion}>
-      <TermsOfUseIcon></TermsOfUseIcon>
-    </AccordionBtn>
-  )
-}
-export default TermsOfUseBtn
